Transaction System
An API capable of managing transactions in real time was created. For this, the DDD architectural paradigm was used for the creation of this program, in addition to using Python with its Locust tool to simulate these transactions and the Proff and Trace profiles to validate the use of memory and different hardware components.